Update the provisioner requires
The rename(s) in chef-provisioning has had follow on changes in the way
the provisioners are loaded. For those Provisioners that have been
updated to be compliant with the /current/ chef-provisioning structure,
upgrade the require lines.

Also revert the blanket renaming for those provisioners that haven't yet
migrated to the new name/structure